Paint R/Y Changes
Updated the Red & Yellow Paint to remove the bugs that existed:

Red paint no longer continuously increases speed to an infinite number.

The code for updating the players movement back to original values have been moved to functions in the Player Blueprint and are being called from Red and Yellow paint Blueprints.